Create cid_geolocate
First version of a rules-based randomized geolocation function. Is only working properly part of the time. Also, based on the output, some adjustments are needed in the distance distribution terms. There might be a need to adjust distance functions when some sections of radiuses away from centroids are shortened -- typically because the geometry abuts a body of water (e.g., New Orleans). The geometry containment test repeats the distance formula but insodoing allows disproportionate weight to placement close to centroid (e.g., distances similar to promixity of a close boundary).