Update cid_geolocate_pgsl
Changed the limit on number of market-geometry (polygon) containment checks from 100 to 300. The 100 limit produced 843 (out of just under 26K) failures to generate a cid_geom value (produced NULLS). Changing it to 300 and re-running the script for the 843 led to all 843 getting assigned a value and thus, eliminating all the NULLs. The main reason I used the more restrictive value was in troubleshooting. It was likely other problems that caused the script to fail fully completing. It is possible, however, that whenever a run the script again for a large number or cids, 300 might be too loose. I'm leaving this note as a reminder in case the possibility has to be addressed in the future.