Create eid_cid_assign
This script uses kmeans clustering of cid geolocations to isolate clusters into which to assign employees (represented by eid -- employee id) to clusters of cids. To be realistic, employees need to be assigned sites they can travel to in reasonable amounts of distance and time. Because of the complexity in which the cid's were clustered, kmeans was a good way to assign eid's to cids. Clustering is by mkt and number of clusters is the number of cids divided by the number of employees in the market.