Update cid_eid_geolocate2
This one improves on previous by geodistributing on a basis proportional to the number of employees in a market, so that employees are assigned roughly the same number of cids and the cids they are assigned are located roughly continuously geographically speaking.  Density of cid/eid assignments reflects density of cids within geography.